+++ crates/app/src/queries/tasks.rs
@@ -56,7 +56,7 @@ fn load_with_regex(
         }
 
         if let (Some(regex), Ok(project_id)) = (&project_regex, task.target.get_project_id())
-            && !regex.is_match(project_id.as_str())
+            && !regex.is_match(project_id)
         {
             continue;
         }